+++ spec/cassandra/cluster/metadata_spec.rb
@@ -22,7 +22,7 @@ require 'spec_helper'
 class FakeSchema < Cassandra::Cluster::Schema
   def add_keyspace(name, replication_class, replication_options)
     replication = Cassandra::Keyspace::Replication.new(replication_class, replication_options)
-    @keyspaces[name] = Cassandra::Keyspace.new(name, true, replication, {}, {}, {}, {})
+    @keyspaces[name] = Cassandra::Keyspace.new(name, true, replication, {}, {}, {}, {}, {})
     self
   end
 end